Settings Config
Added a settings.json file for holding settings data.
Wired up the settings gui with the the settings.json file.
All settings appear to be working properly.

+++ src/main/java/com/monkeystomp/spirelands/graphics/Screen.java
@@ -45,7 +45,7 @@ public class Screen {
    * @param scaleX The horizontal scaling factor to size the final screen.
    * @param scaleY The vertical scaling factor to size the final screen.
    */
-  public Screen(int width, int height, int scaleX, int scaleY) {
+  public Screen(int width, int height, float scaleX, float scaleY) {
     this.width = width;
     this.height = height;
     this.scaleX = scaleX;
